fix cards
