+++ src/app/shared/components/achievement-card/achievement-card.component.html
@@ -3,14 +3,18 @@
       <div class="wrap-for-details" fxLayout="row" fxLayoutAlign="start space-between">
         <img class='image' src='../../../../assets/icons/award.svg'/>
         <div>
-          <h3 class="achievement_title">{{achievementsTitle[achievement.achievementTypeId].name}}</h3>
+          <h3 class="achievement_title">
+            {{achievementsTitle[achievement.achievementTypeId].name}}
+          </h3>
             <h4 class="card_title">Дата</h4> 
             <p class="card_text">{{achievement.achievementDate | date: 'shortDate'}}</p>
-            <ng-container *ngFor="let child of achievement.children">
-              <h4 class="card_title" *ngIf="child">Учасники</h4> 
-              <p class="card_text">
-                {{child.firstName}} {{child.lastName}} {{child.middleName}}
-              </p>
+            <ng-container >
+              <div *ngFor="let child of achievement.children">
+                <h4 class="card_title" *ngIf="child">Учасники</h4> 
+                <p class="card_text">
+                  {{child.firstName}} {{child.lastName}} {{child.middleName}}
+                </p>
+              </div>
             </ng-container>            
             <h4 class="card_title">Опис</h4> 
 
@@ -18,12 +22,13 @@
               <p class="achievement_descr" [ngClass]="{'show-more': showMore}">
                 {{(showMore) ? achievement.title : achievement.title | slice:0:150}}
                 <a *ngIf="achievement.title.length > 70" class='more_btn' 
-                  (click)="showMore = !showMore">{{(!showMore) ? '...Більше' : '...Менше'}}</a>
+                  (click)="showMore = !showMore">{{(!showMore) ? '...Більше' : '...Менше'}}
+                </a>
               </p>  
             </mat-card-content>
         </div> 
         
-        <ng-container *ngIf="auth && getWorkshopIds(workshops | async).includes(workshop.id)">
+        <ng-container *ngIf="getWorkshopIds(workshops | async).includes(workshop.id)">
           <div fxLayout="row" fxLayoutAlign="end" class="actions">
             <div class="actions_btn" [routerLink]="['/create-achievement', achievement.id]">
               <mat-icon class="achievement_icon">edit</mat-icon>
