+++ src/app/shared/components/achievement-card/achievement-card.component.ts
@@ -1,5 +1,14 @@
 import { Component, Input, OnInit } from '@angular/core';
+import { Subject } from '@microsoft/signalr';
+import { Select, Store } from '@ngxs/store';
+import { Observable } from 'rxjs';
+import { AchievementsTitle } from '../../constants/constants';
 import { Achievement } from '../../models/achievement.model';
+import { Workshop } from '../../models/workshop.model';
+import { RegistrationState } from '../../store/registration.state';
+import { GetWorkshopsByProviderId } from '../../store/user.actions';
+import { UserState } from '../../store/user.state';
+import { Provider } from 'src/app/shared/models/provider.model';
 
 @Component({
   selector: 'app-achievement-card',
@@ -7,11 +16,26 @@ import { Achievement } from '../../models/achievement.model';
   styleUrls: ['./achievement-card.component.scss']
 })
 export class AchievementCardComponent implements OnInit {
+  @Select(UserState.workshops) workshops: Observable<Workshop[]>;
+  destroy$: Subject<boolean> = new Subject<boolean>();  
   @Input() achievement: Achievement;
+  @Input() workshop: Workshop;
+  readonly achievementsTitle = AchievementsTitle;
+  showMore = false;
+  auth: boolean;
+  provider: Provider;
 
-  constructor() { }
+  constructor(private store: Store) { }
 
   ngOnInit(): void {
+    this.auth = this.store.selectSnapshot<boolean>(RegistrationState.isAuthorized);
+    this.store.dispatch(new GetWorkshopsByProviderId(this.provider?.id));  
+  }
+
+  getWorkshopIds(arr): string[] {
+    const res = []
+    arr.forEach((i) => res.push(i.workshopId));
+    return res;
   }
 
 }
